# pegasusX/contracts/jwt-core.openapi.yaml
openapi: 3.0.3
info:
  title: PegasusX Human JWT Core API
  version: "1.0.0"
  description: |
    Codegen-ready OpenAPI for the high-traffic human JWT spine (~45 operations).
    Authenticate with `Authorization: Bearer <JWT>` issued by role login/refresh.
    Partner machine API is separate — see [`partner.openapi.yaml`](./partner.openapi.yaml).
    This contract does **not** replace the hand-written `@pegasusx/api-client`; it is the
    substrate for future openapi-typescript / SDK generation. Full ~411-path catalog is out of scope.
servers:
  - url: https://api-ssmr.pegasusx.app
    description: SSMR
security:
  - BearerJWT: []
paths:
...
  /v1/orders/{orderID}:
    get:
      tags: [Orders]
      summary: Get order by ID
      operationId: getOrder
      parameters:
        - name: orderID
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderSummary"
        "404":
          $ref: "#/components/responses/Problem"
...
components:
  securitySchemes:
    BearerJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Human session JWT from role login/refresh
...
  responses:
    Problem:
      description: Problem+JSON error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
  schemas:
    Problem:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
...
    OrderSummary:
      type: object
      properties:
        order_id:
          type: string
        retailer_id:
          type: string
        supplier_id:
          type: string
        status:
          type: string
        total_minor:
          type: integer
        currency:
          type: string
        updated_at:
          type: string
          format: date-time
...